refactor(server): remove lgnodes under nodes; integrate into real Node classes; create internal lgnodes stubs; clean nodes exports (#211)
