+++ apps/server/src/lgnodes/enforceRestriction.lgnode.ts
@@ -1,0 +1,1 @@
+export {}; // lgnodes moved; implementation integrated into nodes/*.node.ts
