--- apps/server/src/nodes/README.txt
@@ -1,1 +1,0 @@
-This directory contains graph nodes.
